requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
idna==2.10
influxdb==5.3.1
isort==5.5.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
numpy==1.19.2
pandas==1.1.2
protobuf==3.13.0
pylint==2.6.0
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyserial==3.4
python-dateutil==2.8.1
pytz==2020.1
requests==2.25.0
six==1.15.0
speedtest-cli==2.1.2
toml==0.10.1
urllib3==1.26.2
wrapt==1.12.1
